adding table preview
